Fold the admin task list into the existing PeriodictaskAdminController
